Created an invoice controller and a Track Controller
The Invoice controller takes in a list of Tracks and the purchasing customer. So if an invoice should be created for an album, you should turn the album into a list of tracks.

The Track controller allows a customer to buy a track and an admin to create, update and delete a Track.

Everyone can view a track